ui/app/partials/configuration.filter.items.php
	ui/app/views/popup.condition.common.php
	ui/include/views/configuration.graph.list.php
	ui/include/views/configuration.httpconf.list.php
	ui/include/views/configuration.triggers.list.php

+++ src/zabbix_server/ha/ha_manager.c
@@ -20,7 +20,7 @@
 #include "ha.h"
 #include "zbxha.h"
 
-#include "db.h"
+#include "zbxdbhigh.h"
 #include "zbxipcservice.h"
 #include "zbxserialize.h"
 #include "zbxthreads.h"
@@ -1667,7 +1667,7 @@ int	zbx_ha_stop(char **error)
 
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s()", __func__);
 
-	if (ZBX_THREAD_ERROR == ha_pid)
+	if (ZBX_THREAD_ERROR == ha_pid || 0 != kill(ha_pid, 0))
 	{
 		ret = SUCCEED;
 		goto out;
@@ -1687,7 +1687,8 @@ int	zbx_ha_stop(char **error)
 		ret = SUCCEED;
 	}
 out:
-	ha_pid = ZBX_THREAD_ERROR;
+	if (SUCCEED == ret)
+		ha_pid = ZBX_THREAD_ERROR;
 
 	zabbix_log(LOG_LEVEL_DEBUG, "End of %s():%s", __func__, zbx_result_string(ret));
 
@@ -1701,9 +1702,12 @@ out:
  ******************************************************************************/
 void	zbx_ha_kill(void)
 {
-	kill(ha_pid, SIGKILL);
-	zbx_thread_wait(ha_pid);
-	ha_pid = ZBX_THREAD_ERROR;
+	if (ZBX_THREAD_ERROR != ha_pid)
+	{
+		kill(ha_pid, SIGKILL);
+		zbx_thread_wait(ha_pid);
+		ha_pid = ZBX_THREAD_ERROR;
+	}
 }
 
 /******************************************************************************
